+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Major changes and resolved issues
 
+## 2023-03-17
+- New page:
+	- Trademark Policy
+
 ## 2023-03-14
 - Update `world.convex/net.clj`